simplified copley comms, removed useless attempt to extract shared behaviour into "Comms" interface, refactored bit/byte conversion into separate project

+++ uk.co.saiman.comms.copley/src/uk/co/saiman/comms/copley/Int16.java
@@ -27,11 +27,11 @@
  */
 package uk.co.saiman.comms.copley;
 
-import uk.co.saiman.comms.Bit;
-import uk.co.saiman.comms.Bits;
+import uk.co.saiman.bytes.Bit;
+import uk.co.saiman.bytes.Bits;
 
 public class Int16 {
 	@Bit(0)
 	@Bits(value = 16)
-	public int value;
+	public short value;
 }
